feat(P-2): renamed definitions imports to types
