+++ src/theme/containers/Footer/cwrap.ts
@@ -1,4 +1,4 @@
-import { TCSSObjProps } from "../../../definitions";
+import { TCSSObjProps } from "../../../types";
 
 export const blocks: TCSSObjProps = {
    core: {
