+++ src/theme/containers/Loader/cwrap.ts
@@ -1,4 +1,4 @@
-import { TCSSObjProps } from "../../../definitions";
+import { TCSSObjProps } from "../../../types";
 import { slideUp } from "./keyframes";
 
 export const blocks: TCSSObjProps = {
